Create Intake program
This is a program I created for the neuropsychological testing facility I work at. It has a tkinker gui that allows the receptionist to ask the potential patient important information to help with efficiency. Afterward, it generates a .docx file to allow us to print it out. This helps the doctors in the office navigate the important information about the patient without needing to decipher messy handwriting. 

This program is actively used in the office. - May 20204